% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adds.R
\name{add_polygenic}
\alias{add_polygenic}
\title{Add polygenic effect}
\usage{
add_polygenic(
  obj,
  W = feature_map$vanraden(),
  G_Polygenic = 1,
  r = NULL,
  scale = FALSE,
  seed = NULL
)
}
\description{
Add polygenic effect
}
