% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_map.R
\docType{data}
\name{feature_map}
\alias{feature_map}
\title{Collect transform functions}
\format{
An object of class \code{list} of length 2.
}
\usage{
feature_map
}
\description{
Collect transform functions
}
\keyword{datasets}
