% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp_simulation.R
\name{snp_simulation}
\alias{snp_simulation}
\title{Simulate SNP data}
\usage{
snp_simulation(
  N = 300,
  M = 600,
  K = 10,
  J = 1,
  p_f = NULL,
  bias = 1,
  variance = 1,
  rho = 0.99,
  p_AA = 0.5,
  p_CC = p_AA,
  sigma_l = 2,
  sigma_lk = NULL,
  sigma_e = 1,
  p_linkage = 0.99,
  p_mate = 1/J,
  seed = NULL
)
}
\description{
Simulate SNP data
}
